feat: Changed filenames around
This commit re-arranges files according to my new vison of things.
It makes clear that the `PolyKerma.*.Interfaces.pas` files will contain all the interfaces for that specific group of classes.
It also makes clear that I want one file per class, without the `I` prefix for the classes that are interfaced.
Since I've found that the `cntrs` unit has some of the lists and hash maps I'm in need, no need to re-invent the wheel, maybe just make them interfaced.