fix: Messages are now well handled
After changing the Proces Messages Thread's Execute method to have the
`ProcessMessage` outside the message get/delete code, I messed up the
deletion. This is now fixed.
Also made sure that all memory was well accounted for and freed.
This also includes the use of the `Post` method, and not the entire
`Dispatcher` as the param for the `Module` creation.